Move registery from init to the files
Registery is getting more complicated so I don't want it all in init anymore.
Now how it works is when the add-on registers, it registers the load handler, and autosave timer, like usual, right? Yes but wait, registery for those is changed. The handler actually re-registers the timer, so that autosave will happen exactly the right amount of time after loading a new file. But then why do I bother manually registering timers on add-on enable? Because enabling the add-on doesn't trigger the load handler, but should still start the autosave timer.