Change some prefs defaults
Autosave interval is now 2 minutes by default, same as blender.
And I've decided to make the datetime format use its original form again, because you're not gonna go back for these on another day anyway. I think how it will be used is to just doodle, and if you think your doodle is worth anything, save as with an actual name.