Improve load handler
The path now uses .resolve() to make sure symlinks and such work.
I no longer to tag_redraw() or reregister the autosave timer, because those wouldn't be useful if it didn't save, and if it did save, would be done by the save handler.
And finally makedirs and save_mainfile are in a try except block that prints the offending path to console.
No need to tag_redraw() when you start a new file and it doesn't save btw, because loading a file does its own tag_redraw() already.