Fix save incremental utility function
I was looking for numbers in the full path, when I should have been looking only for numbers in the stem. Otherwise it could increment a number in a folder, and then try to save there without first making the folder.
Also whoops apparently I forgot the paranthesis for the call to get_increment()