Make sure existing files aren't overwritten
With this new utility function, increment until unique.
It does what save increment did, but over and over until a filename that doesn't exist yet is found.
It replaces most of the code in save incremental, and is added in save datetime in case the user does it multiple times in the same minute, or summer / winter time is doing its thing.