Make path sanitation work on the whole path
Except the anchor, because it contains a colon on windows.
First it resolves the path, to get rid of any ..'s and fix slash orientation, as well as make it absolute.
Then it stores the anchor for later.
Then it removes the anchor from the path.
Then it sanitizes what's left.
And finally it joins it back with the anchor.
Also this is done before increment until unique now, because else it might check if a file exists with invalid chars, and if it doesn't, try to make it with valid chars, which might overwrite a file. Also I'm using the sanitize util from common now.