Use save copy for autosave
This works perfectly but it makes the current blend think that it's saved too, which is not true. That means green icon too, and no prompt when you quit, somewhat misleading. Another problem is that the .autosave file is overwritten from the main file every autosave, which when you load the main file again, you'll lose your work from the .autosave when the autosave timer happens.