Clean up panel code
Seems hops is a function, and I'm guessing mx meant for the code to call this function, so I added the required brackets for that now.
Speaking of brackets, I added a space between PowerSave and (hops) for the label.
Also removed wm since it isn't used, and added some newlines.