Improve powersave function
When you install the addon, the powersave_name field is empty. Now if you happened to be in a file, and immediately tried to use the incremental save feature, it would instead do a datetime save, because the text field was empty. No more! It now checks if you're in a saved file with an empty text field, and updates the powersave_name if necessary. I wanted to set it during addon registery, but apparently bpy.data.filepath isn't accessible at that point. Thanks MindsEye for finding this issue.