Add update function for base_folder
Blender's file browser sets its file name to the last component of the path you opened it with. In the case of a folder, that's the last folder. But at the same time, it puts you inside that folder. This means that if you click browse and then click okay, you'll end up with for example C:\Example\PowerSave\PowerSave, which is obviously no good. I'm hoping this bit of code will prevent 99% of these cases, until the blender devs fix this file browser issue.