Simplify code that draws PowerSave button
No need for HardOps preferences, just check if the operator class name exists in bpy.types.
I wanted to do this because a user couldn't see the PowerSave button, because they had HardOps installed, but a version that was so old it didn't have the hops.powersave operator.
He updated and that fixed it, but I still wanted to make sure that bug never happens again, so this fixes that too.