Add draw and powersave to wm.powersave
PowerLink has utils in wm.powerlink, why shouldn't PowerSave?
Oh and I just noticed the self argument for draw isn't used, and it doesn't make sense for it to be a staticmethod, but I can't change that, because I want newer versions of PowerSave to still work with older versions of PowerLink and visa versa, so it stays.