Add PowerBackup panel code
I replaced the if block for the text with a dictionary, because it's more elegant.
This will matter more once the Power line of products grows further.
The powerbackup_draw function is a copy paste of the powerlink_draw function, with "link" replaced by "backup".
Don't fix it if it ain't broke.