Add integration for PowerManage
This means I added it to the popover tab enum, the popover draw function, and made the powermanage_draw function, which links to store pages that don't exist yet.

The draw function in the window manager only takes the column, because self isn't used anyway.
I wanted to do this for the other draw functions, but can't because it would break compatibility with older versions.
Doesn't hurt to do it here though, even if the inconsistency might seem a bit odd.