Sanitize autosave folder
If it's empty, use the default `//`.
If it's not empty, and doesn't start with `//` (we don't santize relative paths): if the path is absolute, use it (resolved), otherwise, prepend the two forward slashes to make it a relative path (so if the user types just `autosave`, it becomes `//autosave`, but if they type `/test`, it becomes `C:\test` (on Windows)).
Also I'm making the description more human readable (I hope), and adding subtype DIR_PATH so it gets a folder browse button.