Delay autosave if BoxCutter is running
I was supposed to implement this months ago but forgot.
Anyway, if BoxCutter is currently cutting, it is not wise to save, as doing so can cause unexpected things to happen.
Because a cut takes a very short amount of time, I just tell the autosave timer to try again in 1 second.