Adjust autosave_format items
Just changing the names so it's easier to understand without having to waste space on "Autosave Format" in the panel.
Also getting rid of the main description, because I think just using the descriptions from each item makes more sense.
In the future I might want to add Autosave Never or something as an option, instead of having a checkbox for Use Autosave.