Fix error message in verify_version
If the condition is met, path will be None, so it will not have a "name" attribute.
Because this function is only ever called from operators whose poll methods already check for this, the whole if is unnecessary.
But I'm keeping it anyway just in case, and changing the error message to be more generic (so that it wouldn't crash trying to construct the string).