Add Increment Strict option
If enabled, doing an incremental save will only recognize numbers that are prefixed the same way as Increment Format.
For example if a file ends with ".01", but your Increment Format is "_01", it will not change it to ".02", but instead append "_01".
I'm not strict about the number of digits, because I think that would cause issues, with version "_100" for example saving as "_100_01", because 100 has three digits while the preferences have two.