Adjust size of icons
The outlines were smaller for PowerSave and PowerManage compared to PowerLink and PowerBackup.
I made a comparison, and the two people I asked said they preferred the bigger versions of the icons.
So I gave all the icons the bigger outline, moved the lightning bolt 1 pixel to the left, and made the checkmark 2 pixels bigger.